+++ libutil/src/lib.rs
@@ -1,8 +1,7 @@
 //! Library for some useful utilities used by sunflower and it's libraries.
 
 #![cfg_attr(not(test), no_std)]
-#![feature(sync_unsafe_cell, generic_const_exprs)]
-#![allow(incomplete_features)]
+#![feature(sync_unsafe_cell)]
 
 pub use send::{AsBytes, LoadRegisterError, TableDescriptor};
 pub use sync::{ExclusiveMap, InitError, InitLater, UnsafeFlag};